+++ appcontainer/OptionsDialogModel.ts
@@ -8,7 +8,8 @@ import {FormModel} from '@xh/hoist/cmp/form';
 import {AppOptionSpec, HoistModel, managed, XH} from '@xh/hoist/core';
 import {action, computed, makeObservable, observable} from '@xh/hoist/mobx';
 import {assign, mapValues, pickBy} from 'lodash';
-import { resolve } from '../promise/Promise';
+import {isOmitted} from '@xh/hoist/utils/impl';
+import {resolve} from '../promise/Promise';
 import {AppOption} from './AppOption';
 
 /**
@@ -40,7 +41,9 @@ export class OptionsDialogModel extends HoistModel {
     // Setting options
     //-------------------
     setOptions(options: AppOptionSpec[]) {
-        this.options = options.filter(o => !o.omit).map(o => new AppOption(o));
+        this.options = options
+            .filter(o => !isOmitted(o))
+            .map(o => new AppOption(o));
         const fields = this.options.map(o => assign({name: o.name}, o.fieldModel));
         this.formModel = new FormModel({fields, xhImpl: true});
     }