+++ desktop/cmp/contextmenu/ContextMenu.ts
@@ -8,7 +8,7 @@ import {hoistCmp, HoistProps, MenuItem, MenuItemLike} from '@xh/hoist/core';
 import '@xh/hoist/desktop/register';
 import {menu, menuDivider, menuItem} from '@xh/hoist/kit/blueprint';
 import {wait} from '@xh/hoist/promise';
-import {filterConsecutiveMenuSeparators} from '@xh/hoist/utils/impl';
+import {filterConsecutiveMenuSeparators, isOmitted} from '@xh/hoist/utils/impl';
 import {clone, isEmpty, isString} from 'lodash';
 import {isValidElement, ReactElement, ReactNode} from 'react';
 
@@ -55,7 +55,7 @@ function parseItems(items: MenuItemLike[]): ReactNode[] {
     });
 
     return items
-        .filter(it => !isMenuItem(it) || (!it.hidden && !it.omit))
+        .filter(it => !isMenuItem(it) || (!it.hidden && !isOmitted(it)))
         .filter(filterConsecutiveMenuSeparators())
         .map(item => {
             // Process dividers